% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.capscale.R
\name{ds.capscale}
\alias{ds.capscale}
\title{Initiates the capscale of a server-side R object}
\usage{
ds.capscale(formula, data.name, datasources = NULL)
}
\arguments{
\item{formula}{an parameter of unspecified purpose}

\item{data.name}{a character string providing the name of the input object.}

\item{datasources}{a list of DSConnection-class objects obtained after login.}
}
\description{
Uses capscale to create R object on the server-side.
}
\details{
The function create a capscale object of the server-side
}
\author{
Stuart Wheater, DataSHIELD Development Team
}
