% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.capscale_assign.R
\name{ds.capscale_assign}
\alias{ds.capscale_assign}
\title{Initiates the capscale of a server-side R object}
\usage{
ds.capscale_assign(formula, data.name, newobj = NULL, datasources = NULL)
}
\arguments{
\item{formula}{a providing the formula of for processing.}

\item{data.name}{a character string providing the name of the input object.}

\item{newobj}{the name of the new object.}

\item{datasources}{a list of DSConnection-class objects obtained after login.}
}
\description{
Uses capscale to create R object on the server-side.
}
\details{
The function create a capscale object of the server-side
}
\author{
Stuart Wheater, DataSHIELD Development Team
}
