% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.diversity.R
\name{ds.diversity}
\alias{ds.diversity}
\title{Initiates the diversity of a server-side R object}
\usage{
ds.diversity(data.name, datasources = NULL)
}
\arguments{
\item{data.name}{a character string providing the name of the input object.}

\item{datasources}{a list of DSConnection-class objects obtained after login.}
}
\description{
Uses diversity to create R object on the server-side.
}
\details{
The function create a diversity object of the server-side
}
\author{
Stuart Wheater, DataSHIELD Development Team
}
