% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.vegdist.R
\name{ds.vegdist}
\alias{ds.vegdist}
\title{Initiates the vegdist of a server-side R object}
\usage{
ds.vegdist(data.name, datasources = NULL)
}
\arguments{
\item{data.name}{a character string providing the name of the input object.}

\item{datasources}{a list of DSConnection-class objects obtained after login.}
}
\description{
Uses vegdist to create R object on the server-side.
}
\details{
The function create a vegdist object of the server-side
}
\author{
Stuart Wheater, DataSHIELD Development Team
}
